Revert to xpath method
Reverted this back to xpath because the xmlNextJson method is not consistent, first time the results (al least here on my server in the Netherlands) are completely different from what is shown on the actual page. second time and beyond the results are consistent and correct from what is shown on the actual page.